Renamed to Abyssalcreatures and changed some stuff to AutoValue
