+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/Mobs/Hostile/Steve.java
@@ -1,0 +1,19 @@
+package com.offz.spigot.abyssalcreatures.Mobs.Hostile;
+
+import com.offz.spigot.mobzy.Mobs.Behaviours.HitBehaviour;
+import com.offz.spigot.mobzy.Mobs.Types.HostileMob;
+import com.offz.spigot.mobzy.Pathfinders.PathfinderGoalWalkingAnimation;
+import net.minecraft.server.v1_13_R2.World;
+
+public class Steve extends HostileMob implements HitBehaviour {
+    public Steve(World world) {
+        super(world, "Steve");
+        setSize(2F, 7F);
+    }
+
+    @Override
+    public void createPathfinders() {
+        super.createPathfinders();
+        goalSelector.a(0, new PathfinderGoalWalkingAnimation(this, getStaticBuilder().getModelID()));
+    }
+}
