+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/Mobs/Passive/Ashimite.java
@@ -1,0 +1,30 @@
+package com.offz.spigot.abyssalcreatures.Mobs.Passive;
+
+import com.offz.spigot.mobzy.Mobs.Behaviours.HitBehaviour;
+import com.offz.spigot.mobzy.Mobs.Types.PassiveMob;
+import com.offz.spigot.mobzy.Pathfinders.PathfinderGoalWalkingAnimation;
+import net.minecraft.server.v1_13_R2.EntityAgeable;
+import net.minecraft.server.v1_13_R2.EntityHuman;
+import net.minecraft.server.v1_13_R2.World;
+
+public class Ashimite extends PassiveMob implements HitBehaviour {
+    public Ashimite(World world) {
+        super(world, "Ashimite");
+        setSize(2f, 2);
+    }
+
+    @Override
+    public void onRightClick(EntityHuman player) {
+        player.startRiding(this);
+    }
+
+    @Override
+    public void createPathfinders() {
+        super.createPathfinders();
+        goalSelector.a(0, new PathfinderGoalWalkingAnimation(this, getStaticBuilder().getModelID()));
+    }
+
+    public PassiveMob createChild(EntityAgeable entityageable) {
+        return new Ashimite(this.world);
+    }
+}
