+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/Mobs/Passive/Neritantan.java
@@ -1,0 +1,25 @@
+package com.offz.spigot.abyssalcreatures.Mobs.Passive;
+
+import com.offz.spigot.mobzy.Mobs.Behaviours.HitBehaviour;
+import com.offz.spigot.mobzy.Mobs.Types.PassiveMob;
+import com.offz.spigot.mobzy.Pathfinders.PathfinderGoalTemptPitchLock;
+import com.offz.spigot.mobzy.Pathfinders.PathfinderGoalWalkingAnimation;
+import net.minecraft.server.v1_13_R2.EntityAgeable;
+import net.minecraft.server.v1_13_R2.World;
+
+public class Neritantan extends PassiveMob implements HitBehaviour {
+    public Neritantan(World world) {
+        super(world, "Neritantan");
+    }
+
+    @Override
+    public void createPathfinders() {
+        super.createPathfinders();
+        goalSelector.a(0, new PathfinderGoalWalkingAnimation(this, getStaticBuilder().getModelID()));
+        goalSelector.a(4, new PathfinderGoalTemptPitchLock(this, 1.2D, false, getStaticBuilder().getTemptItems()));
+    }
+
+    public PassiveMob createChild(EntityAgeable entityageable) {
+        return new Neritantan(this.world);
+    }
+}
