+++ AbyssalCreatures/src/main/resources/plugin.yml
@@ -1,0 +1,9 @@
+name: AbyssalCreatures
+version: @plugin_version@
+author: Offz
+main: com.offz.spigot.abyssalcreatures.AbyssalCreatures
+api-version: 1.13
+depend: [Mobzy]
+
+prefix: AbyssalCreatures
+description: An implementation of the Mobzy API for the MineInAbyss server
