--- AbyssialCreatures/src/main/java/com/offz/spigot/abyssialcreatures/Mobs/Hostile/Ottobas.java
@@ -1,20 +1,0 @@
-package com.offz.spigot.abyssialcreatures.Mobs.Hostile;
-
-import com.offz.spigot.mobzy.Mobs.Behaviours.HitBehaviour;
-import com.offz.spigot.mobzy.Mobs.Types.HostileMob;
-import com.offz.spigot.mobzy.Pathfinders.PathfinderGoalWalkingAnimation;
-import net.minecraft.server.v1_13_R2.World;
-
-public class Ottobas extends HostileMob implements HitBehaviour {
-    public Ottobas(World world) {
-        super(world, "Ottobas");
-        this.setSize(2F, 3F);
-
-    }
-
-    @Override
-    public void createPathfinders() {
-        super.createPathfinders();
-        goalSelector.a(0, new PathfinderGoalWalkingAnimation(this, getStaticBuilder().getModelID()));
-    }
-}
