--- AbyssialCreatures/src/main/resources/spawns.yml
@@ -1,113 +1,0 @@
-# block-whitelist: if set, the blocks this mob can spawn on | List of Materials | ex. [GRASS_BLOCK, STONE] | default: no whitelist
-# spawn-pos: the type of position for the mob to be spawned in | AIR/GROUND/OVERHANG | ex. AIR | default: GROUND
-config:
-  name: Abyssial Creatures
-  icon: DIAMOND_BLOCK
-regions:
-  - name: Template
-    spawns:
-      - mob: Fuwagi
-        min-amount: 2
-        max-amount: 3
-        min-light: 7
-        radius: 10
-        block-whitelist:
-          - GRASS_BLOCK
-      - mob: Hammerbeak
-        priority: 0.2
-        spawn-pos: AIR
-      - mob: Benikuchinawa
-        priority: 0.05
-        spawn-pos: AIR
-      - mob: Silkfang
-  - name: Orth
-    spawns:
-      - reuse: Template:Fuwagi
-      - mob: Okibo
-        priority: 0.1
-        block-whitelist:
-          - GRASS_BLOCK
-          - STONE_BRICKS
-  - name: Edge of the Abyss
-    spawns:
-      - reuse: Template:Fuwagi
-      - reuse: Template:Hammerbeak
-      - reuse: Template:Benikuchinawa
-        block-whitelist:
-          - STONE
-          - COBBLESTONE
-  - name: Forest of Temptation
-    spawns:
-      - reuse: Template:Fuwagi
-      - reuse: Template:Hammerbeak
-      - reuse: Template:Benikuchinawa
-      - mob: Corpse Weeper
-        priority: 0.3
-        spawn-pos: AIR
-      - reuse: Template:Fuwagi
-        mob: Rohana
-        min-light: 0
-        max-light: 7
-        block-whitelist:
-          - WATER
-      - mob: Tesuchi
-        min-amount: 1
-        max-amount: 4
-        radius: 15
-        block-whitelist:
-          - GRASS_BLOCK
-      - mob: Ottobas
-        min-amount: 1
-        max-amount: 2
-        max-local-group: 2
-        local-group-radius: 128
-        radius: 30
-        block-whitelist:
-          - WATER
-  - name: Inverted Forest
-    spawns:
-      - reuse: Template:Benikuchinawa
-        priority: 0.4
-      - mob: Inbyo
-        min-amount: 4
-        max-amount: 8
-        radius: 5
-        min-y: 140
-        max-light: 7
-  - name: Great Fault
-    spawns:
-      - reuse: Template:Benikuchinawa
-      - reuse: Template:Fuwagi
-        mob: Neritantan
-        min-amount: 3
-        max-amount: 5
-        min-light: 2
-        block-whitelist:
-          - GRASS_BLOCK
-          - ANDESITE
-          - WHITE_CONCRETE
-      - mob: Madokajack
-        priority: 0.3
-        spawn-pos: AIR
-  - name: The Goblet of Giants
-    spawns:
-      - mob: Tamaugachi
-        block-whitelist:
-          - WATER
-        max-local-group: 1
-        local-group-radius: 128
-      - mob: Kuongatari
-        min-amount: 2
-        max-amount: 4
-        radius: 6
-        block-whitelist:
-          - GRASS_BLOCK
-      - reuse: Forest of Temptation:Rohana
-        min-amount: 3
-        max-amount: 5
-      - mob: Kazura
-        min-amount: 2
-        max-amount: 4
-        radius: 6
-        block-whitelist:
-          - WATER
