+++ Mobzy-API/src/main/java/com/offz/spigot/mobzy/Spawning/MobSpawn.java
@@ -257,7 +257,6 @@ public class MobSpawn implements ConfigurationSerializable {
         return maxGap;
     }
 
-
     public int getMaxLocalGroup() {
         return maxLocalGroup;
     }
