+++ gradle.properties
@@ -1,1 +1,1 @@
-plugin_version= 2.3
+plugin_version=2.3
