+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'Mobzy'
 
-include 'Mobzy-API', 'AbyssialCreatures'
+include 'Mobzy-API', 'AbyssalCreatures'
